[client/rest]: refactor light's chain route
